* changes:
  fix(lib/smc): fix MISRA C:2012 Rule 10.1 in system.c
  fix(runtime): fix MISRA C:2012 Rule 10.1 in psci & realm_attest
  fix(runtime): fix MISRA C:2012 Rule 10.1 in host_call.c
  fix(runtime): fix MISRA C:2012 Rule 10.1 in run.c
  fix(runtime): fix MISRA C:2012 Rule 10.1 in rec.c
  fix(lib/arch): fix MISRA C:2012 Rule 10.1 in sysregs.c

+++ runtime/rsi/psci.c
@@ -313,7 +313,7 @@ void handle_psci(struct rec *rec,
 		break;
 	}
 
-	if ((res->action & FLAG_EXIT_TO_HOST) != 0) {
+	if (((unsigned int)res->action & FLAG_EXIT_TO_HOST) != 0U) {
 		rec_exit->exit_reason = RMI_EXIT_PSCI;
 	}
 }
